# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/contact-reports": {
            "get": {
                "tags": [
                    "Contact Reports"
                ],
                "summary": "Get list of contact reports",
                "description": "Display a listing of the resource",
                "operationId": "getContactReportList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Get records in one page default is 10",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by column",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Sort order (asc, desc)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in different column records",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contact Reports"
                ],
                "summary": "Store contact reports",
                "description": "Store a newly created resource in storage",
                "operationId": "storeContactReport",
                "requestBody": {
                    "description": "Pass Json request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject",
                                    "text",
                                    "type",
                                    "priority",
                                    "contactType",
                                    "initiatedBy",
                                    "companyId",
                                    "createdByEmployee",
                                    "categoryId",
                                    "employees",
                                    "companyEmployees",
                                    "assignee",
                                    "contactSources"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "customer/lead"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "contactType": {
                                        "type": "string"
                                    },
                                    "initiatedBy": {
                                        "type": "string",
                                        "example": "customer/docshero"
                                    },
                                    "companyId": {
                                        "type": "string"
                                    },
                                    "createdByEmployee": {
                                        "type": "string"
                                    },
                                    "categoryId": {
                                        "type": "integer"
                                    },
                                    "employees": {
                                        "type": "object",
                                        "example": "[1,3]"
                                    },
                                    "companyEmployees": {
                                        "type": "object",
                                        "example": "[1,3]"
                                    },
                                    "assignee": {
                                        "type": "integer"
                                    },
                                    "contactSources": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "uploadedFiles": {
                                        "type": "object",
                                        "example": "[]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/contact-reports/get-by-resubmit": {
            "get": {
                "tags": [
                    "Contact Reports"
                ],
                "summary": "Get list of contact reports",
                "description": "Display a listing of the resource",
                "operationId": "getContactReportByResumbit",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Get records in one page default is 10",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in different column records",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/contact-reports/company": {
            "get": {
                "tags": [
                    "Contact Reports"
                ],
                "summary": "Get report by company id",
                "description": "Display a listing of the resource againts the specific company",
                "operationId": "getReportByCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of company record",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Bearer token for API authentication",
                "name": "Authorization",
                "in": "header",
                "tokenUrl": "/auth/token"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}